Pull hasNackBeenSentWithCode method into BaseEhrHandler
This way it can be used by other integration tests.
